[Entity Analytics] Remove links to legacy risk scoring docs (#187585)
## Summary

Related to https://github.com/elastic/security-docs/issues/5489. We had
a few places where we incorrectly link to the legacy
user/host risk scoring docs.

In deleting the links I found we had two `RiskSummary` components which
components which I have renamed `RiskSummaryPanel` and
`FlyoutRiskSummary` to make life easier.

<img width="770" alt="Screenshot 2024-07-04 at 15 47 14"
src="https://github.com/elastic/kibana/assets/3315046/20e26a15-4975-4165-9c56-bb328c508bee">

---------

(cherry picked from commit 12bd8fe53ac94e36fc40bad167250390111d122f)
