+++ packages/kbn-doc-links/src/types.ts
@@ -360,8 +360,6 @@ export interface DocLinks {
     readonly ruleUiAdvancedParams: string;
     readonly entityAnalytics: {
       readonly riskScorePrerequisites: string;
-      readonly hostRiskScore: string;
-      readonly userRiskScore: string;
       readonly entityRiskScoring: string;
       readonly assetCriticality: string;
     };
