+++ x-pack/plugins/security_solution/public/entity_analytics/components/entity_analytics_risk_score/translations.ts
@@ -25,14 +25,6 @@ export const VIEW_ALL = i18n.translate(
   }
 );
 
-export const LEARN_MORE = (riskEntity?: RiskScoreEntity) =>
-  i18n.translate('xpack.securitySolution.entityAnalytics.riskDashboard.learnMore', {
-    defaultMessage: 'Learn more about {riskEntity} risk',
-    values: {
-      riskEntity: getRiskEntityTranslation(riskEntity, true),
-    },
-  });
-
 export const LAST_UPDATED = i18n.translate(
   'xpack.securitySolution.entityAnalytics.riskDashboard.lastUpdatedTitle',
   {
