+++ x-pack/plugins/security_solution/public/entity_analytics/components/risk_score/translations.ts
@@ -40,6 +40,13 @@ export const RISK_SCORE_TITLE = (riskEntity: RiskScoreEntity) =>
     },
   });
 
+export const RISK_SCORING_TITLE = i18n.translate(
+  'xpack.securitySolution.riskScore.overview.riskScoringTitle',
+  {
+    defaultMessage: 'Entity Risk Scoring',
+  }
+);
+
 export const ENTITY_RISK_LEVEL = (riskEntity: RiskScoreEntity) =>
   i18n.translate('xpack.securitySolution.entityAnalytics.riskDashboard.riskLevelTitle', {
     defaultMessage: '{riskEntity} risk level',
