+++ x-pack/plugins/security_solution/public/flyout/entity_details/user_right/content.tsx
@@ -12,7 +12,7 @@ import { useIsExperimentalFeatureEnabled } from '../../../common/hooks/use_exper
 import { AssetCriticalityAccordion } from '../../../entity_analytics/components/asset_criticality/asset_criticality_selector';
 
 import { OBSERVED_USER_QUERY_ID } from '../../../explore/users/containers/users/observed_details';
-import { RiskSummary } from '../../../entity_analytics/components/risk_summary_flyout/risk_summary';
+import { FlyoutRiskSummary } from '../../../entity_analytics/components/risk_summary_flyout/risk_summary';
 import type { RiskScoreState } from '../../../entity_analytics/api/hooks/use_risk_score';
 import { ManagedUser } from '../../../timelines/components/side_panel/new_user_detail/managed_user';
 import type { ManagedUserData } from '../../../timelines/components/side_panel/new_user_detail/types';
@@ -58,7 +58,7 @@ export const UserPanelContent = ({
     <FlyoutBody>
       {riskScoreState.isModuleEnabled && riskScoreState.data?.length !== 0 && (
         <>
-          <RiskSummary
+          <FlyoutRiskSummary
             riskScoreData={riskScoreState}
             recalculatingScore={recalculatingScore}
             queryId={USER_PANEL_RISK_SCORE_QUERY_ID}
